provide C bridge to the Gmic library
These two files are part of the official Gmic distribution.
As of 2/2020, the official DEB package fails to link them in.

Moreover, the upstream plugin code has a fallback in case the
decompressing and loading of the gmic standard library fails.
We do not need to install this fallback (gmic_stdlib_gmic.h),
since on Debian systems the package manager ensures proper
installation of the gmic library and the ability to decompress
the embedded gmic standard library image.